+++ components.json
@@ -13,7 +13,7 @@
   "aliases": {
     "components": "@/src/components",
     "utils": "@/src/libs/shadcn-ui/utils",
-    "ui": "@/src/components/ui",
+    "ui": "@/src/components",
     "lib": "@/src/libs",
     "hooks": "@/src/hooks"
   },
