+++ src/app/criar-conta/page.tsx
@@ -1,4 +1,4 @@
-import RegisterForm from "@/src/components/register-form";
+import RegisterForm from "@/src/app/criar-conta/components/register-form";
 
 export default function RegisterScreen() {
   return <RegisterForm />;
