+++ src/app/dashboard/page.tsx
@@ -1,4 +1,4 @@
-import FinancialDashboard from "@/src/components/financial-dashboard";
+import FinancialDashboard from "./components/financial-dashboard";
 
 export default function Home() {
   return (
