+++ src/app/dashboard/services/index.ts
@@ -1,0 +1,3 @@
+export * from "./delete-transaction";
+export * from "./get-transactions";
+export * from "./post-transaction";
