+++ src/app/entrar/page.tsx
@@ -1,4 +1,4 @@
-import LoginForm from "@/src/components/login-form";
+import LoginForm from "@/src/app/entrar/components/login-form";
 
 export default function LoginScreen() {
   return <LoginForm />;
