+++ src/components/date-picker.tsx
@@ -3,13 +3,14 @@ import DateFormatter from "@/src/helpers/date-formatter";
 import { ChevronDownIcon } from "lucide-react";
 import { useState } from "react";
 import { pt } from "react-day-picker/locale";
-import Column from "../utils/column";
-import Flex from "../utils/flex";
-import Show from "../utils/show";
 import { Button } from "./button";
 import { Calendar } from "./calendar";
+
 import { Label } from "./label";
 import { Popover, PopoverContent, PopoverTrigger } from "./popover";
+import Column from "./utils/column";
+import Flex from "./utils/flex";
+import Show from "./utils/show";
 
 interface DatePickerProps {
   value?: Date;
